AAP-5241 Clean up following adding delete endpoints - Move project API functions to project file - Move activation API functions to activation file - Move rule API functions to rule file - Move job API functions to job file
