+++ ansible_events_ui/api/activation.py
@@ -1,14 +1,24 @@
 """Activation API endpoints."""
+import json
 import logging
+from typing import List
 
+import aiodocker.exceptions
 import sqlalchemy as sa
-from fastapi import APIRouter, Depends
+import sqlalchemy.orm
+from fastapi import APIRouter, Depends, status
 from fastapi.exceptions import HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from ansible_events_ui import schemas
+from ansible_events_ui.config import Settings, get_settings
 from ansible_events_ui.db import models
-from ansible_events_ui.db.dependency import get_db_session
+from ansible_events_ui.db.dependency import (
+    get_db_session,
+    get_db_session_factory,
+)
+from ansible_events_ui.managers import updatemanager
+from ansible_events_ui.ruleset import activate_rulesets, inactivate_rulesets
 
 logger = logging.getLogger("ansible_events_ui")
 
@@ -135,3 +145,173 @@ async def update_activation(
     ).one_or_none()
 
     return updated_activation
+
+
+async def read_output(proc, activation_instance_id, db_session_factory):
+    # TODO(cutwater): Replace with FastAPI dependency injections,
+    #   that is available in BackgroundTasks
+    async with db_session_factory() as db:
+        line_number = 0
+        done = False
+        while not done:
+            line = await proc.stdout.readline()
+            if len(line) == 0:
+                done = True
+                continue
+            line = line.decode()
+            await updatemanager.broadcast(
+                f"/activation_instance/{activation_instance_id}",
+                json.dumps(["Stdout", {"stdout": line}]),
+            )
+            query = sa.insert(models.activation_instance_logs).values(
+                line_number=line_number,
+                log=line,
+                activation_instance_id=activation_instance_id,
+            )
+            await db.execute(query)
+            await db.commit()
+            line_number += 1
+
+
+@router.post("/api/activation_instance/")
+async def create_activation_instance(
+    a: schemas.ActivationInstance,
+    db: AsyncSession = Depends(get_db_session),
+    db_session_factory: sqlalchemy.orm.sessionmaker = Depends(
+        get_db_session_factory
+    ),
+    settings: Settings = Depends(get_settings),
+):
+    query = sa.select(models.rulebooks).where(
+        models.rulebooks.c.id == a.rulebook_id
+    )
+    rulebook_row = (await db.execute(query)).first()
+
+    query = sa.select(models.inventories).where(
+        models.inventories.c.id == a.inventory_id
+    )
+    inventory_row = (await db.execute(query)).first()
+
+    query = sa.select(models.extra_vars).where(
+        models.extra_vars.c.id == a.extra_var_id
+    )
+    extra_var_row = (await db.execute(query)).first()
+
+    query = sa.insert(models.activation_instances).values(
+        name=a.name,
+        rulebook_id=a.rulebook_id,
+        inventory_id=a.inventory_id,
+        extra_var_id=a.extra_var_id,
+        working_directory=a.working_directory,
+        execution_environment=a.execution_environment,
+    )
+    result = await db.execute(query)
+    await db.commit()
+    (id_,) = result.inserted_primary_key
+
+    try:
+        await activate_rulesets(
+            settings.deployment_type,
+            id_,
+            a.execution_environment,
+            rulebook_row.rulesets,
+            inventory_row.inventory,
+            extra_var_row.extra_var,
+            a.working_directory,
+            settings.server_name,
+            settings.port,
+            db,
+        )
+    except aiodocker.exceptions.DockerError as e:
+        return HTTPException(status_code=500, detail=str(e))
+
+    return {**a.dict(), "id": id_}
+
+
+@router.post("/api/deactivate/")
+async def deactivate(activation_instance_id: int):
+    await inactivate_rulesets(activation_instance_id)
+    return
+
+
+@router.get("/api/activation_instances/")
+async def list_activation_instances(
+    db: AsyncSession = Depends(get_db_session),
+):
+    query = sa.select(models.activation_instances)
+    result = await db.execute(query)
+    return result.all()
+
+
+@router.get("/api/activation_instance/{activation_instance_id}")
+async def read_activation_instance(
+    activation_instance_id: int, db: AsyncSession = Depends(get_db_session)
+):
+    query = (
+        sa.select(
+            models.activation_instances.c.id,
+            models.activation_instances.c.name,
+            models.rulebooks.c.id.label("ruleset_id"),
+            models.rulebooks.c.name.label("ruleset_name"),
+            models.inventories.c.id.label("inventory_id"),
+            models.inventories.c.name.label("inventory_name"),
+            models.extra_vars.c.id.label("extra_var_id"),
+            models.extra_vars.c.name.label("extra_vars_name"),
+        )
+        .select_from(
+            models.activation_instances.join(models.rulebooks)
+            .join(models.inventories)
+            .join(models.extra_vars)
+        )
+        .where(models.activation_instances.c.id == activation_instance_id)
+    )
+    result = await db.execute(query)
+    return result.first()
+
+
+@router.delete(
+    "/api/activation_instance/{activation_instance_id}",
+    status_code=204,
+    operation_id="delete_activation_instance",
+)
+async def delete_activation_instance(
+    activation_instance_id: int, db: AsyncSession = Depends(get_db_session)
+):
+    query = sa.delete(models.activation_instances).where(
+        models.activation_instances.c.id == activation_instance_id
+    )
+    results = await db.execute(query)
+    if results.rowcount == 0:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+    await db.commit()
+
+
+@router.get(
+    "/api/activation_instance_logs/",
+    response_model=List[schemas.ActivationLog],
+)
+async def list_activation_instance_logs(
+    activation_instance_id: int, db: AsyncSession = Depends(get_db_session)
+):
+    query = (
+        sa.select(models.activation_instance_logs)
+        .where(
+            models.activation_instance_logs.c.activation_instance_id
+            == activation_instance_id
+        )
+        .order_by(models.activation_instance_logs.c.id)
+    )
+    result = await db.execute(query)
+    return result.all()
+
+
+@router.get("/api/activation_instance_job_instances/{activation_instance_id}")
+async def read_activation_instance_job_instances(
+    activation_instance_id: int, db: AsyncSession = Depends(get_db_session)
+):
+    query = sa.select(models.activation_instance_job_instances).where(
+        models.activation_instance_job_instances.c.activation_instance_id
+        == activation_instance_id
+    )
+    result = await db.execute(query)
+    return result.all()
